Make links language agnostic
It doesn't make sense to randomly choose English & Russian versions of these links.  Better to be consistent in 1 of the 2 following ways:
- Since this _is_ the AdGuard wiki's English translation, make all links _explicitly_ point to the English translation
- The server seems to have the ability, that when no language is specified in the KB URL, to automagically choose 1 (depending on browser signalling or user's GEOIP, perhaps?)

This PR is a crude attempt to implement the latter approach.  If it's acceptable, maybe there should be an internal automatic tool to regularly update _all_ KB links this way?